Refactor React and server stream packages
