+++ .changeset/quiet-streams-refactor.md
@@ -1,0 +1,6 @@
+---
+"@anvia/react": patch
+"@anvia/server": patch
+---
+
+Refactor stream helpers into focused internal modules, add coverage-gated tests, and omit stack traces from default streamed server error events.
